Reorganize replayer draft types
